data_pipeline/requirements.txt
pandas==2.2.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0
schedule==1.2.1